Commit casi final

+++ src/utils/cloudStorage.js
@@ -1,9 +1,13 @@
-import { Storage } from "@google-cloud/storage";
 import {storage} from "../config/storage.js"
 import moment from "moment/moment.js";
-async function uploadFile(file) {
+async function uploadFile(file,tabla) {
     const now = moment().format('YYYYMMDD_HHmmss');
-    const bucket = storage.bucket('img-productos');
+    var bucket
+    if(tabla == "productos"){
+      bucket = storage.bucket('img-productos');
+    }else if( tabla == "personas"){
+      bucket = storage.bucket('la-herencia-users-img');
+    }
     const fileName = `${now}_${file.originalname}`;
     const fileUpload = bucket.file(fileName);
     const stream = fileUpload.createWriteStream({
@@ -23,17 +27,23 @@ async function uploadFile(file) {
         stream.end(file.buffer);
     });
   }
-  async function replaceImage(file, nombreImagenAnterior) {
-    const bucket = storage.bucket('img-productos');
+  async function replaceImage(file, nombreImagenAnterior,tabla) {
+    var bucket
+    if(tabla == "productos"){
+      bucket = storage.bucket('img-productos');
+    }else if( tabla == "personas"){
+      bucket = storage.bucket('la-herencia-users-img');
+    }
     // Verificamos si el nombre de la imagen anterior existe en el Cloud Storage
     const blob = await bucket.file(nombreImagenAnterior).get();
     // Si el nombre de la imagen anterior existe en el Cloud Storage, la borramos
-    if (blob) {
+    console.log(nombreImagenAnterior);
+    if (blob && nombreImagenAnterior != "perfil.png") {
       await bucket.file(nombreImagenAnterior).delete();
     }
   
     // Subimos la nueva imagen
-    const nuevaImagenUrl = await uploadFile(file);
+    const nuevaImagenUrl = await uploadFile(file,tabla);
   
     // Obtenemos el nombre del archivo de la nueva imagen
     const nombreImagenNueva = nuevaImagenUrl
